+++ tests/testthat/test-compare.r
@@ -15,3 +15,16 @@ test_that("equal if both missing or both the same", {
   expect_false(compare("ABC", NA_character_)$equal)
   expect_false(compare(NA_character_, "ABC")$equal)
 })
+
+test_that("equal if both missing or both the same (multiple values)", {
+  expect_true(compare(c("ABC", NA), c("ABC", NA))$equal)
+
+  expect_false(compare(c(NA, NA), c("ABC", NA))$equal)
+  expect_false(compare(c("AB", NA), c("ABC", NA))$equal)
+  expect_false(compare(c("AB", "AB"), c("ABC", "AB"))$equal)
+})
+
+test_that("computes correct number of mismatches", {
+  comp <- compare(c("a","b","c"), c("c", "d", "e"))
+  expect_match(comp$message, "3 string mismatches")
+})